A synthetic dataset contains variables for eGFR calculation.

A synthetic dataset contains variables for eGFR calculation for 1000 adults and 1000 children.

Usage

ckd.data

Format

A data frame with 2000 rows (1000 adults and 1000 children/young adults) and 12 variables:

cr

Serum creatinine, micromol/L

cys

Serum cystatin C, mg/L

age

Age, years

sex

Sex

ethnicity

Ethnicity

height

Height, cm

category

Indication on whether the generated data refer to adults or children

Source

Synthetic dataset was generated based on two publications:

adults: Lamb EJ, Barratt J, Brettell EA et al. Accuracy of glomerular filtration rate estimation using creatinine and cystatin C for identifying and monitoring moderate chronic kidney disease: the eGFR-C study. Health Technol Assess 2024;28(35), doi:10.3310/HYHN1078.

children/young adults: Pierce CB, Muñoz A, Ng DK et al. Age- and sex-dependent clinical equations to estimate glomerular filtration rates in children and young adults with chronic kidney disease. Kidney International. 2021;99(4):948–956, doi:10.1016/j.kint.2020.10.047.

Details

Citation: Bikbov B. kidney.epi: Kidney-Related Functions for Clinical and Epidemiological Research. Scientific-Tools.Org, https://Scientific-Tools.Org. DOI: 10.32614/CRAN.package.kidney.epi