Calculate eGFR by the EKFC cystatin-based equation

Calculate eGFR by the EKFC cystatin-based equation

Usage

egfr.ekfc.cys(cystatin, age, cystatin_units = "mg/L", max_age = 100)

Arguments

cystatin

Numeric vector. Serum cystatin, could be expressed in "mg/L" or "nanomol/L". Units of measurement should be defined in variable cystatin_units (if not defined explicitly by user, the default value is "mg/L").

age

Numeric vector. Age, in years.

cystatin_units

Character string. Units in which serum cystatin is expressed. Could be one of the following: "mg/L" or "nanomol/L"

max_age

Numeric. Maximal age suitable for the equation application, in years. By default is 100 years, but change this value in case you would like to apply equation to older persons.

Value

numeric eGFR expressed in ml/min/1.73m2.

Details

Calculate estimated glomerular filtration rate (eGFR) by the EKFC cystatin-based equation.

Reference to the equation: Pottel H, Björk J, Rule AD, et al. Cystatin C–based equation to estimate GFR without the inclusion of race and sex. N Engl J Med. 2023;388:333-343 doi: 10.1056/NEJMoa22037.

Citation: Bikbov B. kidney.epi: Kidney-Related Functions for Clinical and Epidemiological Research. Scientific-Tools.Org, https://Scientific-Tools.Org. DOI: 10.32614/CRAN.package.kidney.epi

Author

Programming: Boris Bikbov https://www.linkedin.com/in/boris-bikbov.

Examples

# for a single patient
egfr.ekfc.cys (cystatin = 0.8, age = 60)
#> [1] 90.5
# for a dataset - see vignettes for details
# egfr.ekfc.cys (cystatin = dta$cys, age = dta$age)