Calculate eGFR by the Full age spectrum (FAS) creatinine-based equation

Calculate eGFR by the Full age spectrum (FAS) creatinine-based equation

Usage

egfr.fas.cr(
  creatinine,
  age,
  sex,
  creatinine_units = "micromol/l",
  label_sex_male = c("Male", 1),
  label_sex_female = c("Female", 0),
  max_age = 100
)

Arguments

creatinine

Numeric vector. Serum creatinine, could be expressed in "micromol/L", "mmol/L" or "mg/dL". Units of measurement should be defined in variable creatinine_units (if not defined explicitly by user, the default value is "micromol/L").

age

Numeric vector. Age, in years.

sex

Vector. The value of variable refers to the parameters label_sex_male and label_sex_female.

creatinine_units

Character string. Units in which serum creatinne is expressed. Could be one of the following: "micromol/L", "mmol/L" or "mg/dL".

label_sex_male

List. Label(s) for definition(s) of male sex.

label_sex_female

List. Label(s) for definition(s) of female sex.

max_age

Numeric. Maximal age suitable for the equation application, in years. By default is 100 years, but change this value in case you would like to apply equation to older persons.

Value

numeric eGFR expressed in ml/min/1.73m2.

Details

Calculate estimated glomerular filtration rate (eGFR) by the Full age spectrum (FAS) creatinine-based equation.

Reference to the equation: Pottel H, Hoste L, Dubourg L et al. An estimating glomerular filtration rate equation for the full age spectrum. Nephrol Dial Transplant 2016; 31:798–806 doi:10.1093/ndt/gfv454.

Citation: Bikbov B. kidney.epi: Kidney-Related Functions for Clinical and Epidemiological Research. Scientific-Tools.Org, https://Scientific-Tools.Org. DOI: 10.32614/CRAN.package.kidney.epi

Author

Programming: Boris Bikbov https://www.linkedin.com/in/boris-bikbov.

Examples

# for a single patient
egfr.fas.cr (creatinine = 1.4, age = 60, sex = "Male", 
  creatinine_units = "mg/dl")
#> [1] 54.48
# for a dataset - see vignettes for details
# egfr.fas.cr (creatinine = dta$scr, age = dta$age, sex = dta$sex, 
#  creatinine_units = "mg/dl")